+++ scylla/tests/integration/statements/mod.rs
@@ -5,3 +5,4 @@ mod silent_prepare_query;
 mod skip_metadata_optimization;
 mod statement;
 mod timestamps;
+mod transparent_reprepare;
